feat: 컴포넌트 및 컨텍스트 분리
