+++ src/contexts/NotificationContext.tsx
@@ -1,0 +1,57 @@
+import React, { createContext, useContext, useState } from "react";
+import { Notification } from "../type/type";
+
+interface NotificationContextType {
+  notifications: Notification[];
+  addNotification: (message: string, type: Notification["type"]) => void;
+  removeNotification: (id: number) => void;
+}
+
+const NotificationContext = createContext<NotificationContextType | undefined>(
+  undefined,
+);
+
+export const useNotificationContext = () => {
+  const context = useContext(NotificationContext);
+  if (context === undefined) {
+    throw new Error(
+      "useNotification must be used within an NotificationContext",
+    );
+  }
+  return context;
+};
+
+export const NotificationContextProvider = ({
+  children,
+}: {
+  children: React.ReactNode;
+}) => {
+  const [notifications, setNotifications] = useState<Notification[]>([]);
+
+  const addNotification = (message: string, type: Notification["type"]) => {
+    const newNotification: Notification = {
+      id: Date.now(),
+      message,
+      type,
+    };
+    setNotifications((prev) => [...prev, newNotification]);
+  };
+
+  const removeNotification = (id: number) => {
+    setNotifications((prev) =>
+      prev.filter((notification) => notification.id !== id),
+    );
+  };
+
+  const notificationContextValue: NotificationContextType = {
+    notifications,
+    addNotification,
+    removeNotification,
+  };
+
+  return (
+    <NotificationContext.Provider value={notificationContextValue}>
+      {children}
+    </NotificationContext.Provider>
+  );
+};
